Delete article.json
Articles are integrated into item names while mining. Originally, they would have `<info::1:3>` or such in front, but this is merged with the Articles list to produce `<info::a :the >`.

On the one hand, this makes item data more readable. On the other hand, it makes editing articles harder as each item has its own copy of the actual string. On the gripping hand, the articles aren't expected to change. If they do, just re-mine the data.

Just need to remember that that's a non-breaking space (U+00A0), not a regular space (U+0020).